go.mod: require Go 1.26+
