+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/nspcc-dev/dbft
 
-go 1.25
+go 1.26
 
 require (
 	github.com/stretchr/testify v1.11.1
